Add ComposerIO implementation with types
The latest version of `composer/composer` uses type declarations, which
is not backwards compatible with PHP 5.6.

See https://github.com/wp-cli/package-command/issues/172
